+++ src/main/resources/db/migration/procedures/flow/workQueue/R__get_work_queue_metrics.sql
@@ -25,8 +25,8 @@ CREATE OR REPLACE FUNCTION flow.get_work_queue_metrics(p_work_queue_type_id bigi
 as
 $$
 declare
-  v_long_window                       bigint;
-  v_short_window                      bigint;
+  v_long_window                       integer;
+  v_short_window                      integer;
   v_long_window_duration_type_id      bigint;
   v_short_window_duration_type_id     bigint;
   v_expected_cycle                    bigint;
